requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.4
djangorestframework==3.15.1
fonttools==4.34.4
gunicorn==22.0.0
psycopg2==2.2.9
psycopg2-binary==2.9.9
pycparser==2.21
pyparsing==3.0.9
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15